resolvers added for mutations in typedef, client side mutation and query coded added
