+++ client/src/utils/queries.ts
@@ -1,0 +1,20 @@
+import { gql } from '@apollo/client';
+
+export const GET_ME = gql`
+    query me {
+        me {
+            _id
+            username
+            email
+            bookCount
+            savedBooks {
+                _id
+                bookId
+                title
+                authors
+                description
+                image
+                link
+            }
+        }
+    }`;
