+++ server/src/models/Book.ts
@@ -1,40 +1,49 @@
-import { Schema, type Document } from 'mongoose';
+import { Schema, model, Document } from 'mongoose';
 
-export interface BookDocument extends Document {
-  bookId: string;
-  title: string;
-  authors: string[];
-  description: string;
-  image: string;
-  link: string;
+export interface IBook extends Document {
+    bookId: string;
+    title: string;
+    authors: string[];
+    description: string;
+    image: string;
+    link: string;
 }
 
-// This is a subdocument schema, it won't become its own model but we'll use it as the schema for the User's `savedBooks` array in User.js
-const bookSchema = new Schema<BookDocument>({
-  // saved book id from GoogleBooks
-  bookId: {
-    type: String,
-    required: true,
-  },
-  authors: [
+const bookSchema = new Schema<IBook>(
     {
-      type: String,
-    },
-  ],
-  description: {
-    type: String,
-    required: true,
-  },
-  image: {
-    type: String,
-  },
-  link: {
-    type: String,
-  },
-  title: {
-    type: String,
-    required: true,
-  },
-});
+        bookId: {
+            type: String,
+            required: true,
+        },
+        title: {
+            type: String,
+            required: true,
+        },
+        authors: [
+            {
+                type: String,
+            },
+        ],
+        description: {
+            type: String,
+            required: true,
+        },
+        image: {
+            type: String,
+        },
+        link: {
+            type: String,
+        },
+    }
+    // TODO: uncomment for settings for schema if needed
+    // ,
+    // {
+    //     timestamps: true,
+    //     toJSON: { getters: true },
+    //     toObject: { getters: true },
+    // }
+);
 
-export default bookSchema;
+const Book = model<IBook>('Book', bookSchema);
+
+export default Book;
