+++ server/src/models/User.ts
@@ -1,68 +1,68 @@
-import { Schema, model, type Document } from 'mongoose';
+import { Schema, model, Document } from 'mongoose';
 import bcrypt from 'bcrypt';
 
-// import schema from Book.js
-import bookSchema from './Book.js';
-import type { BookDocument } from './Book.js';
-
-export interface UserDocument extends Document {
-  id: string;
-  username: string;
-  email: string;
-  password: string;
-  savedBooks: BookDocument[];
-  isCorrectPassword(password: string): Promise<boolean>;
-  bookCount: number;
+interface IUser extends Document {
+    username: string;
+    email: string;
+    password: string;
+    savedBooks: Schema.Types.ObjectId[];
+    isCorrectPassword(password: string): Promise<boolean>;
+    bookCount: number;
 }
 
-const userSchema = new Schema<UserDocument>(
-  {
-    username: {
-      type: String,
-      required: true,
-      unique: true,
-    },
-    email: {
-      type: String,
-      required: true,
-      unique: true,
-      match: [/.+@.+\..+/, 'Must use a valid email address'],
-    },
-    password: {
-      type: String,
-      required: true,
-    },
-    // set savedBooks to be an array of data that adheres to the bookSchema
-    savedBooks: [bookSchema],
-  },
-  // set this to use virtual below
-  {
-    toJSON: {
-      virtuals: true,
-    },
-  }
+const userSchema = new Schema<IUser>(
+    {
+        username: {
+            type: String,
+            required: true,
+            unique: true,
+            trim: true,
+        },
+        email: {
+            type: String,
+            required: true,
+            unique: true,
+            match: [/.+@.+\..+/, 'Must match an email address!'],
+        },
+        password: {
+            type: String,
+            required: true,
+            minlength: 8,
+        },
+        savedBooks: [
+            {
+                type: Schema.Types.ObjectId,
+                ref: 'Book',
+            },
+        ],
+        bookCount: {
+            type: Number,
+            default: 0,
+        },
+    }
+    // TODO: uncomment for settings for schema if needed
+    // ,
+    // {
+    //     timestamps: true,
+    //     toJSON: { getters: true },
+    //     toObject: { getters: true },
+    // }
 );
 
-// hash user password
-userSchema.pre('save', async function (next) {
-  if (this.isNew || this.isModified('password')) {
-    const saltRounds = 10;
-    this.password = await bcrypt.hash(this.password, saltRounds);
-  }
+// pre hook to hash the password before saving a new user or updating password field on a user
+userSchema.pre<IUser>('save', async function (next) {
+    if (this.isNew || this.isModified('password')) {
+        const saltRounds = 10;
+        this.password = await bcrypt.hash(this.password, saltRounds);
+    }
 
-  next();
+    next();
 });
 
-// custom method to compare and validate password for logging in
-userSchema.methods.isCorrectPassword = async function (password: string) {
-  return await bcrypt.compare(password, this.password);
+userSchema.methods.isCorrectPassword = async function (password: string): Promise<boolean> {
+    return bcrypt.compare(password, this.password);
 };
 
-// when we query a user, we'll also get another field called `bookCount` with the number of saved books we have
-userSchema.virtual('bookCount').get(function () {
-  return this.savedBooks.length;
-});
-
-const User = model<UserDocument>('User', userSchema);
+const User = model<IUser>('User', userSchema);
 
 export default User;
