+++ server/src/models/index.ts
@@ -1,3 +1,4 @@
-import User from './User';
+import Book from './Book.js';
+import User from './User.js';
 
-export default { User };
+export { Book, User };
