+++ server/src/server.ts
@@ -1,21 +1,45 @@
 import express from 'express';
 import path from 'node:path';
-import db from './config/connection.js';
-import routes from './routes/index.js';
+import type { Request, Response } from 'express';
+import db from './config/connection.js'
+import { ApolloServer } from '@apollo/server';// Note: Import from @apollo/server-express
+import { expressMiddleware } from '@apollo/server/express4';
+import { typeDefs, resolvers } from './schemas/index.js';
+import { authenticateToken } from './utils/auth.js';
 
-const app = express();
-const PORT = process.env.PORT || 3001;
-
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-
-// if we're in production, serve client/build as static assets
-if (process.env.NODE_ENV === 'production') {
-  app.use(express.static(path.join(__dirname, '../client/build')));
-}
-
-app.use(routes);
-
-db.once('open', () => {
-  app.listen(PORT, () => console.log(`🌍 Now listening on localhost:${PORT}`));
+const server = new ApolloServer({
+    typeDefs,
+    resolvers
 });
+
+const startApolloServer = async () => {
+    await server.start();
+    await db();
+
+    const PORT = process.env.PORT || 3001;
+    const app = express();
+
+    app.use(express.urlencoded({ extended: false }));
+    app.use(express.json());
+
+    app.use('/graphql', expressMiddleware(server as any,
+        {
+            context: authenticateToken as any
+        }
+    ));
+
+    if (process.env.NODE_ENV === 'production') {
+        app.use(express.static(path.join(__dirname, '../client/dist')));
+
+        app.get('*', (_req: Request, res: Response) => {
+            res.sendFile(path.join(__dirname, '../client/dist/index.html'));
+        });
+    }
+
+    app.listen(PORT, () => {
+        console.log(`API server running on port ${PORT}!`);
+        console.log(`Use GraphQL at http://localhost:${PORT}/graphql`);
+    });
+};
+
+startApolloServer();
