--- server/src/types/express/index.d.ts
@@ -1,8 +1,0 @@
-declare namespace Express {
-  interface Request {
-    user: {
-      _id: unknown;
-      username: string;
-    };
-  }
-}
