--- serverNew/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/server.js",
-  "type": "module",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/server.js",
-    "watch": "nodemon dist/server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@apollo/server": "^4.11.3",
-    "bcrypt": "^4.0.1",
-    "brcypt": "^1.0.1",
-    "express": "^4.19.2",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^8.0.0"
-  },
-  "devDependencies": {
-    "@types/bcrypt": "^5.0.2",
-    "@types/express": "^4.17.21",
-    "@types/jsonwebtoken": "^9.0.6",
-    "@types/node": "^20.14.8",
-    "dotenv": "^16.4.5",
-    "nodemon": "^3.1.0",
-    "typescript": "^5.4.5"
-  }
-}
