--- serverNew/src/models/index.ts
@@ -1,4 +1,0 @@
-import Book from './Book.js';
-import User from './User.js';
-
-export { Book, User };
