--- serverNew/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    /* Modules */
-    "rootDir": "src" /* Specify the root folder within your source files. */,
-
-    /* Emit */
-    "outDir": "dist" /* Specify an output folder for all emitted files. */
-    // "noEmitOnError": true /* Disable emitting files if any type checking errors are reported. */
-  },
-  "include": ["src"],
-  "typeRoots": ["./node_modules/@types", "./src/types"]
-}
